tests/wpt/tests/html/canvas/element/fill-and-stroke-styles/2d.gradient.radial.transform.2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.gradient.radial.transform.2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.gradient.radial.transform.2</h1>
<p class="desc">Radial gradient coordinates are relative to the coordinate space at the time of filling</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Radial gradient coordinates are relative to the coordinate space at the time of filling");
_addTest(function(canvas, ctx) {

  ctx.translate(100, 0);
  var g = ctx.createRadialGradient(0, 0, 0, 0, 0, 11.2);
  g.addColorStop(0, '#0f0');
  g.addColorStop(0.5, '#0f0');
  g.addColorStop(0.51, '#f00');
  g.addColorStop(1, '#f00');
  ctx.fillStyle = g;
  ctx.translate(-50, 25);
  ctx.scale(10, 10);
  ctx.fillRect(-5, -2.5, 10, 5);
  _assertPixel(canvas, 25,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 75,25, 0,255,0,255);

});
</script>
